UI: Refactor core to be more maintainable
Change-type: patch
