--- lib/ui/core/constants.js
@@ -1,7 +1,0 @@
-
-/*
- * Copyright (C) Balena.io - All Rights Reserved
- * Unauthorized copying of this file, via any medium is strictly prohibited.
- * Proprietary and confidential.
- */
-exports.STORAGE_KEY = 'jellyfish_store'
