+++ lib/ui/core/index.js
@@ -1,85 +1,88 @@
-
 /*
  * Copyright (C) Balena.io - All Rights Reserved
  * Unauthorized copying of this file, via any medium is strictly prohibited.
  * Proprietary and confidential.
  */
+
 /* global process */
 /* eslint-disable no-process-env */
-const Bluebird = require('bluebird')
-const localForage = require('localforage')
-const _ = require('lodash')
-const {
+import localForage from 'localforage'
+import * as _ from 'lodash'
+import {
 	Analytics
-} = require('../services/analytics')
-const helpers = require('../services/helpers')
-const urlManager = require('../services/url-manager')
-const common = require('./common')
-const constants = require('./constants')
-const {
-	sdk
-} = require('./sdk')
-const store = require('./store')
+} from '../services/analytics'
+import helpers from '../services/helpers'
+import UrlManager from '../services/url-manager'
+import {
+	sdk as SDK
+} from './sdk'
+import {
+	setupStore
+} from './store'
+
+export const sdk = SDK
+
 const ANALYTICS_TOKEN = process.env.MIXPANEL_TOKEN_UI
-exports.store = store.createJellyfishStore()
-exports.analytics = new Analytics({
+const STORAGE_KEY = 'jellyfish_store'
+
+export const analytics = new Analytics({
 	token: ANALYTICS_TOKEN
 })
-const load = () => {
-	return Bluebird.try(() => {
-		helpers.debug('LOADING STATE FROM STORAGE')
-		return localForage.getItem(constants.STORAGE_KEY)
 
-		// TODO abstract this logic to something more redux-esque
-			.then((state) => {
-				if (state) {
-				// Remove notifications
-					_.set(state, [ 'core', 'notifications' ], [])
+const bundle = setupStore({
+	analytics,
+	sdk,
+	storageKey: STORAGE_KEY
+})
 
-					// Ensure that the stored state has a safe structure buy merging it with
-					// the default state. This helps gaurd against situations where the
-					// defaultstate changes or localStorage becomes corrupted.
-					// Additionally, 'status' is always set back to 'initializing', so that the
-					// session is re-checked on load, and the UI bootstrapping process
-					// functions in the correct order
-					const defaultState = common.getDefaultState()
+export const selectors = bundle.selectors
+export const store = bundle.store
+export const actionCreators = bundle.actionCreators
 
-					// Remove unknown top level keys from stored state
-					_.forEach(_.keys(state), (key) => {
-						if (!_.has(defaultState, key)) {
-							Reflect.deleteProperty(state, key)
-						}
+const manager = new UrlManager({
+	actionCreators,
+	selectors,
+	store
+})
+
+localForage.getItem(STORAGE_KEY)
+	.then(async (state) => {
+		if (state) {
+			// Remove notifications
+			_.set(state, [ 'core', 'notifications' ], [])
+
+			// Ensure that the stored state has a safe structure buy merging it with
+			// the default state. This helps gaurd against situations where the
+			// defaultstate changes or localStorage becomes corrupted.
+			// Additionally, 'status' is always set back to 'initializing', so that the
+			// session is re-checked on load, and the UI bootstrapping process
+			// functions in the correct order
+			store.dispatch(
+				actionCreators.setState(
+					_.merge(store.getState().core, state.core, {
+						status: 'initializing'
 					})
-					exports.store.dispatch({
-						type: store.actions.SET_STATE,
-						value: _.merge(defaultState.core, state.core, {
-							status: 'initializing'
-						})
-					})
-				}
-			})
-	})
-}
-load()
-	.then(async () => {
-		const token = store.selectors.getSessionToken(exports.store.getState())
+				)
+			)
+		}
+
+		const token = selectors.getSessionToken(store.getState())
+
 		if (token) {
 			helpers.debug('FOUND STORED SESSION TOKEN, CHECKING AUTHORIZATION')
-			await exports.store.dispatch(store.actionCreators.loginWithToken(token))
+			await store.dispatch(actionCreators.loginWithToken(token))
 		} else {
-			await exports.store.dispatch(store.actionCreators.setStatus('unauthorized'))
+			await store.dispatch(actionCreators.setStatus('unauthorized'))
 		}
 
 		// Only try and load the URL route if the app is authorized
-		if (store.selectors.getStatus(exports.store.getState()) === 'authorized') {
-			urlManager.setChannelsFromPath()
+		if (selectors.getStatus(store.getState()) === 'authorized') {
+			manager.setChannelsFromPath()
 		}
 
-		exports.store.subscribe(() => {
-			return urlManager.setPathFromState(exports.store.getState())
+		store.subscribe(() => {
+			return manager.setPathFromState(store.getState())
 		})
-
-		return null
 	})
+
 window.sdk = sdk
-exports.sdk = sdk
