+++ lib/ui/core/store/actions.js
@@ -1,0 +1,37 @@
+/*
+ * Copyright (C) Balena.io - All Rights Reserved
+ * Unauthorized copying of this file, via any medium is strictly prohibited.
+ * Proprietary and confidential.
+ */
+
+export default {
+	SET_STATUS: 'SET_STATUS',
+	SET_STATE: 'SET_STATE',
+	SET_TYPES: 'SET_TYPES',
+	SET_ORGS: 'SET_ORGS',
+	SET_ACCOUNTS: 'SET_ACCOUNTS',
+	SET_ALL_USERS: 'SET_ALL_USERS',
+	UPDATE_CHANNEL: 'UPDATE_CHANNEL',
+	ADD_CHANNEL: 'ADD_CHANNEL',
+	REMOVE_CHANNEL: 'REMOVE_CHANNEL',
+	SET_CHANNELS: 'SET_CHANNELS',
+	SET_AUTHTOKEN: 'SET_AUTHTOKEN',
+	LOGOUT: 'LOGOUT',
+	SET_USER: 'SET_USER',
+	ADD_NOTIFICATION: 'ADD_NOTIFICATION',
+	REMOVE_NOTIFICATION: 'REMOVE_NOTIFICATION',
+	ADD_VIEW_NOTICE: 'ADD_VIEW_NOTICE',
+	REMOVE_VIEW_NOTICE: 'REMOVE_VIEW_NOTICE',
+	SET_CONFIG: 'SET_CONFIG',
+	SET_SEND_COMMAND: 'SET_SEND_COMMAND',
+	SET_UI_STATE: 'SET_UI_STATE',
+
+	// View specific actions
+	STREAM_VIEW: 'STREAM_VIEW',
+	SET_VIEW_LENS: 'SET_VIEW_LENS',
+	SET_VIEW_DATA: 'SET_VIEW_DATA',
+	UPSERT_VIEW_DATA_ITEM: 'UPSERT_VIEW_DATA_ITEM',
+	APPEND_VIEW_DATA_ITEM: 'APPEND_VIEW_DATA_ITEM',
+	REMOVE_VIEW_DATA_ITEM: 'REMOVE_VIEW_DATA_ITEM',
+	SAVE_SUBSCRIPTION: 'SAVE_SUBSCRIPTION'
+}
