+++ lib/ui/lens/Thread.jsx
@@ -17,7 +17,10 @@ const rendition = require('rendition')
 const CardActions = require('../components/CardActions')
 const CardField = require('../components/CardField').default
 const Tag = require('../components/Tag')
-const store = require('../core/store')
+const {
+	actionCreators,
+	selectors
+} = require('../core')
 const helpers = require('../services/helpers')
 const Timeline = require('./Timeline')
 const CloseButton = require('../shame/CloseButton')
@@ -124,13 +127,19 @@ class Base extends React.Component {
 }
 const mapStateToProps = (state) => {
 	return {
-		allUsers: store.selectors.getAllUsers(state),
-		types: store.selectors.getTypes(state)
+		allUsers: selectors.getAllUsers(state),
+		types: selectors.getTypes(state)
 	}
 }
 const mapDispatchToProps = (dispatch) => {
 	return {
-		actions: redux.bindActionCreators(store.actionCreators, dispatch)
+		actions: redux.bindActionCreators(
+			_.pick(actionCreators, [
+				'addChannel',
+				'removeChannel'
+			]),
+			dispatch
+		)
 	}
 }
 exports.Renderer = connect(mapStateToProps, mapDispatchToProps)(Base)
