+++ lib/ui/services/helpers.js
@@ -249,32 +249,6 @@ exports.getUserIdsByPrefix = (prefix, source, users) => {
 		.value()
 }
 
-/**
- * @summary Convert a string into a 32bit hashcode
- *
- * @param {String} input - The input source to hash
- *
- * @returns {Number} - A 32bit integer
- */
-exports.hashCode = (input) => {
-	let hash = 0
-	let iteration = 0
-	let character = ''
-	if (input.length === 0) {
-		return hash
-	}
-	for (iteration; iteration < input.length; iteration++) {
-		character = input.charCodeAt(iteration)
-
-		// eslint-disable-next-line no-bitwise
-		hash = ((hash << 5) - hash) + character
-
-		// Convert to 32bit integer
-		// eslint-disable-next-line no-bitwise
-		hash |= 0
-	}
-	return hash
-}
 exports.getObjectValues = (input) => {
 	if (_.isPlainObject(input)) {
 		const result = _.map(input, (value) => {
