+++ lib/ui/services/link.js
@@ -4,13 +4,17 @@
  * Unauthorized copying of this file, via any medium is strictly prohibited.
  * Proprietary and confidential.
  */
-const core = require('../core')
-const store = require('../core/store')
+const {
+	actionCreators,
+	analytics,
+	store,
+	sdk
+} = require('../core')
 
 exports.createLink = (fromCard, toCard, verb, options = {}) => {
-	return core.sdk.card.link(fromCard, toCard, verb)
+	return sdk.card.link(fromCard, toCard, verb)
 		.tap(() => {
-			core.analytics.track('element.create', {
+			analytics.track('element.create', {
 				element: {
 					type: 'link'
 				}
@@ -18,10 +22,10 @@ exports.createLink = (fromCard, toCard, verb, options = {}) => {
 		})
 		.tap(() => {
 			if (!options.skipSuccessMessage) {
-				core.store.dispatch(store.actionCreators.addNotification('success', 'Created new link'))
+				store.dispatch(actionCreators.addNotification('success', 'Created new link'))
 			}
 		})
 		.catch((error) => {
-			core.store.dispatch(store.actionCreators.addNotification('danger', error.message))
+			store.dispatch(actionCreators.addNotification('danger', error.message))
 		})
 }
