+++ lib/ui/services/notifications.js
@@ -6,37 +6,48 @@
  */
 const howler = require('howler')
 const _ = require('lodash')
-const core = require('../core')
-const store = require('../core/store')
+const {
+	actionCreators,
+	store
+} = require('../core')
 const helpers = require('./helpers')
+
 const TIMEOUT = 10 * 1000
+
 const sound = new howler.Howl({
 	src: '/audio/dustyroom_cartoon_bubble_pop.mp3'
 })
+
 let canUseNotifications = Notification.permission === 'granted'
+
 if (Notification && Notification.permission !== 'denied') {
 	Notification.requestPermission((status) => {
 		// Status is "granted", if accepted by user
 		canUseNotifications = status === 'granted'
 	})
 }
+
 exports.createNotification = (title, body, target) => {
 	if (!canUseNotifications) {
 		return
 	}
+
 	const notice = new Notification(title, {
 		body,
 		icon: '/icons/jellyfish.png'
 	})
+
 	sound.play()
+
 	const timeout = setTimeout(() => {
 		return notice.close()
 	}, TIMEOUT)
+
 	notice.onclick = () => {
-		core.store.dispatch(store.actionCreators.addChannel(helpers.createChannel({
+		store.dispatch(actionCreators.addChannel(helpers.createChannel({
 			cardType: 'view',
 			target,
-			parentChannel: _.get(core.store.getState(), [ 'channels', '0', 'id' ])
+			parentChannel: _.get(store.getState(), [ 'channels', '0', 'id' ])
 		})))
 		clearTimeout(timeout)
 		notice.close()
