+++ lib/ui/services/store-helpers.js
@@ -5,13 +5,15 @@
  * Proprietary and confidential.
  */
 const _ = require('lodash')
-const core = require('../core')
-const store = require('../core/store')
+const {
+	store,
+	selectors
+} = require('../core')
 
 exports.getActor = (id) => {
-	const state = core.store.getState()
-	const allUsers = store.selectors.getAllUsers(state)
-	const accounts = store.selectors.getAccounts(state)
+	const state = store.getState()
+	const allUsers = selectors.getAllUsers(state)
+	const accounts = selectors.getAccounts(state)
 	const actor = _.find(allUsers, {
 		id
 	})
